Add scheduled pump analysis with initial historical data analysis
Implements a scheduler to run pump analysis every 12 hours and adds initial 7-day historical analysis via `scheduler.py`.

+++ pump-analysis/main.py
@@ -781,7 +781,7 @@ class PumpAnalysisSystem:
         # Create data directory
         os.makedirs('pump_data', exist_ok=True)
         
-    def run_analysis(self, days_back: int = 7, max_symbols: int = 30):
+    def run_analysis(self, days_back: float = 7, max_symbols: int = 30):
         """
         Run complete pump analysis
